wip: styles are starting to come together
